second-seminar/src/main/java/com/server/dosopt/secondseminar/service/MemberService.java
package com.server.dosopt.secondseminar.service;

import com.server.dosopt.secondseminar.domain.SOPT;
import com.server.dosopt.secondseminar.dto.request.MemberCreateRequest;
import com.server.dosopt.secondseminar.dto.request.MemberProfileUpdateRequest;
import com.server.dosopt.secondseminar.dto.response.MemberGetResponse;
import com.server.dosopt.secondseminar.domain.entity.Member;
import com.server.dosopt.secondseminar.repository.MemberJpaRepository;
import jakarta.persistence.EntityNotFoundException;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;
import java.util.stream.Collectors;

@Service
@RequiredArgsConstructor
@Transactional(readOnly = true)
public class MemberService {

    private final MemberJpaRepository memberJpaRepository;

    public MemberGetResponse getMemberByIdV1(Long id) {
        Member member = memberJpaRepository.findById(id).get();
        return MemberGetResponse.of(member);
    }

    public MemberGetResponse getMemberByIdV2(Long id) {
        return MemberGetResponse.of(memberJpaRepository.findById(id).orElseThrow(
                () -> new EntityNotFoundException("존재하지 않는 회원입니다.")));
    }

    public MemberGetResponse getMemberByIdV3(Long id) {
        return MemberGetResponse.of(memberJpaRepository.findByIdOrThrow(id));
    }

    /*public MemberGetResponse getMemberByNickname(String nickname) {
        try {
            return MemberGetResponse.of(memberJpaRepository.findByNickname(nickname));
        } catch (Exception e) {
            throw new IllegalStateException("존재하지 않는 회원입니다.");
        }
    }*/

    public List<MemberGetResponse> getMembers() {
        return memberJpaRepository.findAll()
                .stream()
                .map(MemberGetResponse::of)
                .collect(Collectors.toList());
    }

    @Transactional
    public String create(MemberCreateRequest request) {
        Member member =  memberJpaRepository.save(Member.builder()
                .name(request.name())
                .nickname(request.nickname())
                .age(request.age())
                .sopt(request.sopt())
                .build());
        return member.getId().toString();
    }

    @Transactional
    public void updateSOPT(Long memberId, MemberProfileUpdateRequest request) {
        Member member = memberJpaRepository.findByIdOrThrow(memberId);
        member.updateSOPT(new SOPT(request.generation(), request.part()));
    }

    @Transactional
    public void deleteMember(Long memberId) {
        Member member = memberJpaRepository.findByIdOrThrow(memberId);
        memberJpaRepository.delete(member);
    }
}
